Fix circular logic
