+++ src/nn/variation/architecture_adaptation.py
@@ -1,6 +1,7 @@
 import torch
 import torch.nn as nn
 import logging
+import math
 
 from src.nn.variation.utils import add_specific_node
 
@@ -76,3 +77,89 @@ class ReshapeModule(nn.Module):
     
     def forward(self, x):
         return x.reshape(-1, *self.target_size)
+
+def adapt_node_shape(graph, node, current_size, target_size, target_user=None, try_linear_adapter=True):
+    """
+    Adapts a node's output shape to match a target size using repetition, adaptive pooling or circular padding.
+    
+    Args:
+        graph: The FX graph
+        node: The node whose shape needs to be adapted
+        current_size: Current size of the node's output, no batch dimension
+        target_size: Desired size of the node's output, no batch dimension
+        target_user: Optional specific node that should use the adapted output. If None, all users will be updated.
+        try_linear_adapter: If True, try using a single linear layer instead of flatten->adapt->unflatten pattern
+    Returns:
+        graph: The modified graph
+        adapted_node: The node after shape adaptation
+    """
+    # Convert current_size and target_size to tuples if they are not already
+    current_size = tuple(current_size)
+    target_size = tuple(target_size)
+    
+    if current_size == target_size:
+        return graph, node
+    
+    current_dims = len(current_size)
+    target_dims = len(target_size)
+    
+    # Handle 1D to 1D case directly
+    if current_dims == 1 and target_dims == 1:
+        if try_linear_adapter:
+            return add_specific_node(
+                graph,
+                node,
+                nn.Linear(current_size[0], target_size[0]),
+                target_user=target_user
+            )
+        return _adapt_tensor_size(graph, node, current_size[0], target_size[0], target_user=target_user)
+    
+    # Calculate total elements
+    current_total = math.prod(current_size)
+    target_total = math.prod(target_size)
+    
+    # If total elements are the same, just reshape and return
+    if current_total == target_total:
+        return add_specific_node(
+            graph,
+            node,
+            ReshapeModule(target_size),
+            target_user=target_user
+        )
+    
+    if try_linear_adapter and current_size[:-1] == target_size[:-1]:  # Only use linear adapter if all but last dims are the same
+        return add_specific_node(
+            graph,
+            node,
+            nn.Linear(current_size[-1], target_size[-1]),
+            target_user=target_user
+        )
+    
+    # Step 1: Flatten if starting from multi-dimensional (2+:1 or 2+:2+)
+    if current_dims > 1:
+        graph, node = add_specific_node(
+            graph, 
+            node, 
+            nn.Flatten(start_dim=1, end_dim=-1),
+            target_user=target_user
+        )
+    
+    # Step 2: Adapt tensor size (total elements differ, so this is always needed)
+    graph, node = _adapt_tensor_size(
+        graph, 
+        node, 
+        current_total, 
+        target_total, 
+        target_user=target_user
+    )
+    
+    # Step 3: Unflatten if ending with multi-dimensional (1:2+ or 2+:2+)
+    if target_dims > 1:
+        graph, node = add_specific_node(
+            graph, 
+            node, 
+            nn.Unflatten(dim=1, unflattened_size=target_size),
+            target_user=target_user
+        )
+    
+    return graph, node
