+++ src/nn/variation/utils.py
@@ -1,4 +1,3 @@
-import math
 import logging
 
 import torch
@@ -7,7 +6,6 @@ import torch.fx
 from torch.fx.passes.shape_prop import ShapeProp
 
 from ..individual_graph_module import NeuralNetworkIndividualGraphModule
-from .architecture_adaptation import ReshapeModule, _adapt_tensor_size
 
 
 def get_unique_name(graph, base_name: str) -> str:
@@ -131,93 +129,6 @@ def add_skip_connection(graph, second_node, first_node, torch_function=torch.add
 
     return graph, new_node
 
-
-def adapt_node_shape(graph, node, current_size, target_size, target_user=None, try_linear_adapter=True):
-    """
-    Adapts a node's output shape to match a target size using repetition, adaptive pooling or circular padding.
-    
-    Args:
-        graph: The FX graph
-        node: The node whose shape needs to be adapted
-        current_size: Current size of the node's output, no batch dimension
-        target_size: Desired size of the node's output, no batch dimension
-        target_user: Optional specific node that should use the adapted output. If None, all users will be updated.
-        try_linear_adapter: If True, try using a single linear layer instead of flatten->adapt->unflatten pattern
-    Returns:
-        graph: The modified graph
-        adapted_node: The node after shape adaptation
-    """
-    # Convert current_size and target_size to tuples if they are not already
-    current_size = tuple(current_size)
-    target_size = tuple(target_size)
-    
-    if current_size == target_size:
-        return graph, node
-    
-    current_dims = len(current_size)
-    target_dims = len(target_size)
-    
-    # Handle 1D to 1D case directly
-    if current_dims == 1 and target_dims == 1:
-        if try_linear_adapter:
-            return add_specific_node(
-                graph,
-                node,
-                nn.Linear(current_size[0], target_size[0]),
-                target_user=target_user
-            )
-        return _adapt_tensor_size(graph, node, current_size[0], target_size[0], target_user=target_user)
-    
-    # Calculate total elements
-    current_total = math.prod(current_size)
-    target_total = math.prod(target_size)
-    
-    # If total elements are the same, just reshape and return
-    if current_total == target_total:
-        return add_specific_node(
-            graph,
-            node,
-            ReshapeModule(target_size),
-            target_user=target_user
-        )
-    
-    if try_linear_adapter and current_size[:-1] == target_size[:-1]:  # Only use linear adapter if all but last dims are the same
-        return add_specific_node(
-            graph,
-            node,
-            nn.Linear(current_size[-1], target_size[-1]),
-            target_user=target_user
-        )
-    
-    # Step 1: Flatten if starting from multi-dimensional (2+:1 or 2+:2+)
-    if current_dims > 1:
-        graph, node = add_specific_node(
-            graph, 
-            node, 
-            nn.Flatten(start_dim=1, end_dim=-1),
-            target_user=target_user
-        )
-    
-    # Step 2: Adapt tensor size (total elements differ, so this is always needed)
-    graph, node = _adapt_tensor_size(
-        graph, 
-        node, 
-        current_total, 
-        target_total, 
-        target_user=target_user
-    )
-    
-    # Step 3: Unflatten if ending with multi-dimensional (1:2+ or 2+:2+)
-    if target_dims > 1:
-        graph, node = add_specific_node(
-            graph, 
-            node, 
-            nn.Unflatten(dim=1, unflattened_size=target_size),
-            target_user=target_user
-        )
-    
-    return graph, node
-
 def add_branch_nodes(graph: NeuralNetworkIndividualGraphModule, reference_node, branch1_module, branch2_module):
     """
     Adds two branch nodes in parallel after the reference node and connects them with a skip connection.
@@ -228,8 +139,13 @@ def add_branch_nodes(graph: NeuralNetworkIndividualGraphModule, reference_node, 
         branch1_module: The module for the first branch
         branch2_module: The module for the second branch
     Returns:
-        graph: The modified graph
-        new_node: The combined node after the branches
+        tuple: (graph, new_node, branch1_node, branch2_node, branch1_shape, branch2_shape)
+            - graph: The modified graph
+            - new_node: The combined node after the branches
+            - branch1_node: The first branch node (may need adaptation)
+            - branch2_node: The second branch node (may need adaptation)
+            - branch1_shape: The shape of branch1_node's output
+            - branch2_shape: The shape of branch2_node's output
     """
     # Generate unique names for the branch nodes
     branch1_name = get_unique_name(graph, "branch1")
@@ -258,27 +174,12 @@ def add_branch_nodes(graph: NeuralNetworkIndividualGraphModule, reference_node, 
     # Infer the shapes of the branch nodes from the metadata, pass through get_feature_dims to remove batch dimension
     branch1_shape = get_feature_dims(branch1_node.meta['tensor_meta'].shape)
     branch2_shape = get_feature_dims(branch2_node.meta['tensor_meta'].shape)
-    
-    # Initialize variables to track the final nodes to use in skip connection
-    final_branch1_node = branch1_node
-    final_branch2_node = branch2_node
-    
-    # Adapt branch nodes if needed to ensure they have compatible shapes
-    if branch1_shape != branch2_shape:    
-        # Adapt first branch
-        graph, final_branch1_node = adapt_node_shape(graph, branch1_node, branch1_shape, branch2_shape)
 
-    # Run shape propagation to update metadata for the branch nodes
-    ShapeProp(graph).propagate(graph.example_input)
-
-    # Get the inferred output of the skip connection from the shape propagation
-    skip_connection_output_shape = final_branch2_node.meta['tensor_meta'].shape
-
-    # Create a skip connection between the two branch nodes using the final adapted nodes
-    with graph.graph.inserting_after(final_branch2_node):
+    # Create a skip connection between the two branch nodes
+    with graph.graph.inserting_after(branch2_node):
         new_node = graph.graph.call_function(
             torch.add,
-            args=(final_branch1_node, final_branch2_node),
+            args=(branch1_node, branch2_node),
         )
     
     # Update connections - replace all uses of reference_node with new_node
@@ -287,7 +188,7 @@ def add_branch_nodes(graph: NeuralNetworkIndividualGraphModule, reference_node, 
     branch1_node.args = (reference_node,)
     branch2_node.args = (reference_node,)
 
-    return graph, new_node, skip_connection_output_shape
+    return graph, new_node, branch1_node, branch2_node, branch1_shape, branch2_shape
 
 def get_feature_dims(shape):
     """
